AppOpsService: Add MODE_ASK support to AppOps.
Add support for new mode(MODE_ASK) in AppOpsService to show a permission
dialog box to user to confirm user permission before allowing or ignoring
that operation.
All strict operations (defined in AppOpsManager) are going to be in
MODE_ASK by default.
Operations will be moved to MODE_ALLOWED or MODE_IGNORED according to
user's choice.

Change-Id: I1314125a2b8be558e422e4a9eea0ff066c21bf94
Signed-off-by: Alvin Francis <nivlafx@gmail.com>

# Conflicts:
#	core/java/android/app/AppOpsManager.java
#	core/res/res/values/config.xml
#	core/res/res/values/strings.xml